Use custom town name in building names (#648)
* Use custom town name in building names

* Implement changes
