+++ lib/index.ts
@@ -175,9 +175,10 @@ export * from './town/getTownWelfare'
 export * from './town/politicsWeightedRoll'
 export * from './town/setTownMaterialProbability'
 export * from './town/townData'
+export * from './town/replaceTownName'
 export * from './town/roads'
 export * from './town/townRender'
 export * from './town/updateTownSocioPolitics'
 export * from './town/townDemographics'
-
+export * from './src/gpt3'
 export * from './types'
