+++ src/Start/Start.twee
@@ -26,3 +26,4 @@
       passage: 'Welcome'
     })>>
 <</if>>
+<<print lib.getGPTResponse('Hello there! I am a')>>
