+++ src/Town/js/createTownName.ts
@@ -23,39 +23,7 @@ export const createTownName = function (town: TownBasics | Town) {
   const driftName = lib.linguisticDrift(name)
 
   // casting as Town because replaceTownName tests for Town properties.
-  replaceTownName(town as Town, driftName)
+  lib.replaceTownName(town as Town, driftName)
 
   return driftName
-}
-
-const replaceTownName = (town: Town, driftName: string) => {
-  if (town?.buildings && town.generated === 'full') replaceBuildingNames(town, driftName)
-  if (town?.factions && town.generated === 'full') replaceFactionNames(town, driftName)
-}
-
-const replaceFactionNames = (town: Town, driftName: string) => {
-  console.log(town.name, driftName)
-  // Replace existing names of factions if they reference town name
-  for (const faction of Object.values(town.factions)) {
-    town.factions[faction.key].name = town.factions[
-      faction.key
-    ].name.replace(town.name, driftName)
-    town.factions[faction.key].tippyDescription = town.factions[
-      faction.key
-    ].tippyDescription.replace(town.name, driftName)
-  }
-}
-
-const replaceBuildingNames = (town: Town, driftName: string) => {
-  console.log(town.name, driftName)
-  // Replace existing names of buildings if they reference town name
-  town.buildings.forEach((building) => {
-    if (building.name) { building.name = building.name.replace(town.name, driftName) }
-    if (building.tippyDescription) {
-      building.tippyDescription = building.tippyDescription.replace(
-        town.name,
-        driftName
-      )
-    }
-  })
 }
